# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Pheromone Management API
  version: "0.1.0"
  description: |
    REST API for the Pheromone digital-twin platform management server.

    All endpoints under `/api/v1/` (except `/api/v1/auth/login` and
    `/api/v1/health`) require a Bearer token obtained via
    `POST /api/v1/auth/login`.

    Real-time agent and twin state updates are available via the
    Server-Sent Events stream at `GET /api/v1/events`.

    Kubernetes liveness and readiness probes are at `/healthz` and
    `/readyz` respectively (no authentication required).
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/agents:
    get:
      summary: List all registered agents
      operationId: listAgents
      tags: [agents]
      responses:
        "200":
          description: Array of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/twins:
    get:
      summary: List all digital twins
      operationId: listTwins
      tags: [twins]
      responses:
        "200":
          description: Array of twins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Twin"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/groups:
    get:
      summary: List all groups
      operationId: listGroups
      tags: [groups]
      responses:
        "200":
          description: Array of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      summary: Create a new group
      operationId: createGroup
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/changesets:
    get:
      summary: List all changesets
      operationId: listChangesets
      tags: [changesets]
      responses:
        "200":
          description: Array of changesets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Changeset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: HMAC-SHA256-signed token obtained from POST /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    MethodNotAllowed:
      description: HTTP method not supported for this endpoint
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    AgentIdentity:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          example: tls-cert
        value:
          type: string
        label:
          type: string
    Agent:
      type: object
      required: [id, name, type, status, last_seen, registered_at]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [os, workload]
        status:
          type: string
          enum: [online, offline, stale]
        hostname:
          type: string
        twin_ids:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        identities:
          type: array
          items:
            $ref: "#/components/schemas/AgentIdentity"
        last_seen:
          type: string
          format: date-time
        registered_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
    Twin:
      type: object
      required: [id, name, type, state, config_version, last_heartbeat, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [os, workload]
        state:
          type: string
          enum: [active, stale, inactive]
        agent_id:
          type: string
        config_version:
          type: integer
        last_heartbeat:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
        actual_state:
          type: object
          additionalProperties: true
        desired_state:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FieldChange:
      type: object
      required: [field, old_value, new_value]
      properties:
        field:
          type: string
        old_value:
          type: string
        new_value:
          type: string
    Changeset:
      type: object
      required: [id, twin_id, agent_id, type, status, changes, created_at]
      properties:
        id:
          type: string
        twin_id:
          type: string
        agent_id:
          type: string
        type:
          type: string
          enum: [config, state]
        status:
          type: string
          enum: [pending, applied, rejected]
        changes:
          type: array
          items:
            $ref: "#/components/schemas/FieldChange"
        approved_by:
          type: string
        applied_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Group:
      type: object
      required: [id, name, type, members, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [agents, twins, infrastructure]
        description:
          type: string
        members:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [agents, twins, infrastructure]
        description:
          type: string
        members:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: HTTP status text
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Human-readable error detail